requirements.txt
pygame
python-chess
stockfish
pytest